Unnið í HoleService og Hole.java
