+++ src/main/java/hbv501g/Persistence/Repositories/HoleRepository.java
@@ -1,0 +1,9 @@
+package hbv501g.Persistence.Repositories;
+
+import hbv501g.Persistence.Entities.Hole;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+public interface HoleRepository extends JpaRepository<Hole,Long> {
+
+}
